Refactor: SearchBar refactoring
