+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "AutoPreview.serverURL": "http://localhost:3000/"
+}
