+++ src/Pureplate/Header/Header.jsx
@@ -1,11 +1,11 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import styles from "./Pureplate.module.css";
-import SearchBar from "../Search/SearchBar.js";
-import Attributes from "../Attributes/Attributes.js";
-import Profile from "../Profile/Profile.jsx";
-import BookmarkIcon from "../Bookmark/BookmarkIcon.jsx";
-import logo_icon from "../assets/Icons/logo_icon.png";
+import styles from "./Header.module.css";
+import SearchBar from "../../Search/SearchBar.js";
+import Attributes from "../../Attributes/Attributes.js";
+import Profile from "../../Profile/Profile.jsx";
+import BookmarkIcon from "../../Bookmark/BookmarkIcon.jsx";
+import logo_icon from "../../assets/Icons/logo_icon.png";
 
 function Header({ isRestModalOpen, bookmarkToggle }) {
   return (
