+++ src/Pureplate/Header/Header.module.css
@@ -1,0 +1,42 @@
+
+.header {
+    border-style: solid;
+    border-color: transparent;
+    border-width: 0.0625rem;
+    padding: 0.625rem 0.625rem 0.625rem 0.625rem;
+    display: flex;
+    flex-direction: row;
+    gap: 0rem;
+    align-items: center;
+    justify-content: flex-start;
+    width: 100%;
+    height: 5rem;
+    position: absolute;
+    left: 0rem;
+    top: 0rem;
+    box-shadow: 0rem 0.0625rem 1.875rem 0rem rgba(0, 0, 0, 0.06);
+  }
+  
+  @media (max-width: 1000px) {
+    .header {
+      padding: 0.625rem 0.625rem 0.625rem 0.625rem;
+    }
+  }
+  
+  .purePlateIcon {
+    width: 60px;
+    height: 60px;
+    position: relative;
+    align-items: center;
+    object-fit: contain;
+    padding-right: 10px;
+    padding-top: 5px;
+  }
+  
+  @media (max-width: 600px) {
+    .purePlateIcon {
+      min-width: 40px;
+      min-height: 40px;
+    }
+  }
+
