+++ src/Pureplate/Pureplate.jsx
@@ -3,7 +3,7 @@ import { useParams, useNavigate, Outlet } from "react-router-dom";
 import styles from "./Pureplate.module.css";
 import MapNaverCur from "../Map/Map.js";
 import RestaurantModal from "./RestaurantModal.jsx";
-import Header from "./Header.jsx";
+import Header from "./Header/Header.jsx";
 import { useAuth } from "../AuthContext.jsx";
 
 function Pureplate() {
