+++ src/Pureplate/Pureplate.module.css
@@ -212,42 +212,6 @@
   bottom: 47.95%;
   top: 50.68%;
 }
-
-.header {
-  border-style: solid;
-  border-color: transparent;
-  border-width: 0.0625rem;
-  padding: 0.625rem 0.625rem 0.625rem 0.625rem;
-  display: flex;
-  flex-direction: row;
-  gap: 0rem;
-  align-items: center;
-  justify-content: flex-start;
-  width: 100%;
-  height: 5rem;
-  position: absolute;
-  left: 0rem;
-  top: 0rem;
-  box-shadow: 0rem 0.0625rem 1.875rem 0rem rgba(0, 0, 0, 0.06);
-}
-
-@media (max-width: 1000px) {
-  .header {
-    padding: 0.625rem 0.625rem 0.625rem 0.625rem;
-  }
-}
-
-.purePlateIcon {
-  flex-shrink: 1;
-  width: 70%;
-  height: 70%;
-  min-width: 50%;
-  min-height: 50%;
-  position: relative;
-  align-items: center;
-  object-fit: contain;
-}
-
 .attributes {
   padding: 0.625rem;
   display: flex;
