+++ src/Search/Dropdown.module.css
@@ -1,0 +1,60 @@
+@font-face {
+  font-family: 'NanumGothic';
+  src: url('../assets/fonts/NanumGothic.otf') format('truetype')
+}
+
+.dropDownBox {
+    background: rgba(255, 255, 255, 0.96);
+    border-radius: 0 0 0.5rem 0.5rem;
+    display: flex;
+    flex-direction: column;
+    gap: 0.625rem;
+    box-shadow: 0 0.25rem 0.25rem rgba(0, 0, 0, 0.25);
+    overflow: hidden;
+    list-style: none;
+    width: 17.2rem;
+    position: absolute;
+    padding: 0.5rem;
+    padding-bottom: 1rem;
+  }
+  
+  .dropDownItem {
+    padding-top: 1rem;
+    width: 100%;
+    z-index: 4;
+    letter-spacing: 2px;
+    display: flex;
+    align-items: center;
+    padding-left: 10px;
+    font-family: 'NanumGothic' sans-serif;
+  }
+  
+  .dropDownItem.selected {
+    background-color: #edf5f5;
+  }
+  
+  .union {
+    display: flex;
+    gap: 0.625rem;
+    align-items: center;
+    width: 1.4rem;
+    fill: #AAAAAA;
+  }
+  
+  .recentSearch12 {
+    color: rgba(98, 98, 98, 0.89);
+    text-align: left;
+    font-family: "NanumGothic", sans-serif;
+    font-size: 1rem;
+    display: flex;
+    align-items: center;
+  }
+  
+  .searchHistory {
+    padding: 0 0.7rem;
+    display: flex;
+    gap: 0.3525rem;
+    align-items: center;
+    flex: 1;
+  }
+
