+++ src/Search/SearchBar.module.css
@@ -1,143 +1,61 @@
-/* Entire Search Area*/
+@font-face {
+  font-family: 'NanumGothic';
+  src: url('../assets/fonts/NanumGothic.otf') format('truetype')
+}
+
 .searchBar {
   display: flex;
-  flex-direction: row;
-  gap: 0rem;
   align-items: center;
-  justify-content: flex-start;
-  flex-grow: 0;
-  flex-shrink: 1;
-  flex-basis: 24rem;
+  width: 22rem;
   min-width: 15rem;
   height: 3.2rem;
-  position: relative;
   padding-right: 15px;
 }
 
+.searchLayer {
+  height: 100%;
+  width: 100%;
+}
+
 .searchForm {
-  background: rgba(255, 255, 255, 0.96);
+  background: transparent;
+  flex-direction: column;
+  height: 100%;
+}
+
+.searchInput {
+  background-color: white;
   border-style: solid;
   border-color: #d1d1d167;
   border-width: 0.0525rem;
-  display: flex;
-  flex-direction: column;
-  gap: 0.625rem;
-  align-items: flex-start;
-  justify-content: flex-start;
-  flex: 1;
+  border-radius: 0.7rem 0rem 0rem 0.7rem;
+  color: rgba(0, 0, 0, 0.8);
+  font-family: "NanumGothic", sans-serif;
+  font-size: 1.2rem;
   height: 100%;
-  position: relative;
-  color: rgba(0, 0, 0, 0.5);
-  font-family: "AnekBangla-ExtraLight", sans-serif;
-  font-size: 1.3rem;
-  font-weight: 200;
-  border-radius: 0.5rem 0rem 0rem 0.5rem;
-}
-
-/* Actual Input occured here */
-.searchInput {
-  color: rgba(0, 0, 0, 0.5);
-  font-family: "AnekBangla-ExtraLight", sans-serif;
-  font-size: 1.4rem;
-  font-weight: 200;
-  line-height: 149.8%;
-  position: relative;
   width: 100%;
-  height: 100%;
   display: flex;
-  align-items: center;
   justify-content: center;
   padding-left: 20px;
 }
 
 .searchInput:focus {
-  border: 0px solid #cfcfcf7a;
-  outline: none; /* 기본 outline 제거 (선택 사항) */
+  border: none;
+  outline: none;
+}
+
+/* Input 창에 값이 있을 때 달라지는 스타일 */
+.searchInput.hasValue {
+  border-radius: 0.5rem 0rem 0rem 0.0rem;
+  box-shadow: 0 0.25rem 0.25rem rgba(0, 0, 0, 0.25);
 }
 
 .searchButton {
-  flex-shrink: 0;
   width: 3.875rem;
   height: 100%;
-  position: relative;
-  overflow: hidden;
   background: rgba(74, 124, 89, 0.96);
-  border-radius: 0rem 0.5rem 0.5rem 0rem;
+  border-radius: 0rem 0.6rem 0.6rem 0rem;
   cursor: pointer;
-}
-
-.dropDownBox {
-  z-index: 3;
-  background: rgba(255, 255, 255, 0.96);
-  border-radius: 0rem 0rem 0.5rem 0.5rem;
-  display: flex;
-  flex-direction: column;
-  gap: 0.625rem;
-  align-items: center;
-  justify-content: flex-start;
-  box-shadow: 0rem 0.25rem 0.25rem 0rem rgba(0, 0, 0, 0.25);
-  overflow: hidden;
-  list-style-type: none;
-  width: 100%;
-  position: absolute;
-  top: 104%; /* searchForm의 높이에 맞춰 조정 */
-}
-
-.dropDownItem {
-  padding: 10px 8px;
-  width: 100%;
-  z-index: 4;
-  letter-spacing: 2px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: flex-start;
-  align-self: stretch;
-  flex-shrink: 0;
-  position: relative;
-  font-weight: 400;
-}
-
-.dropDownItem.selected {
-  background-color: #edf5f5;
-}
-
-.union {
-  display: flex;
-  flex-direction: row;
-  gap: 0.625rem;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-  width: 1.4rem;
-  height: auto;
-  position: relative;
-  overflow: visible;
-  fill: #AAAAAA;
-}
-
-.recentSearch12 {
-  color: rgba(53, 53, 53, 0.75);
-  text-align: left;
-  font-family: "NotoSansKr-Regular", sans-serif;
-  font-size: 1.1275362968444824rem;
-  font-weight: 400;
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-}
-
-.searchHistory {
-  padding: 0rem 0.7rem;
-  display: flex;
-  flex-direction: row;
-  gap: 0.3525rem;
-  align-items: center;
-  justify-content: flex-start;
-  align-self: stretch;
-  flex: 1;
-  position: relative;
 }
 
 .searchIcon {
@@ -147,12 +65,3 @@
   justify-content: center;
   padding: 13px;
 }
-
-@media (max-width: 1050px) {
-  .searchBar {
-    width: auto;
-    height: 90%;
-  }
-}
-
-
